Move email token confirmation to verify module
